% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_data.R
\docType{class}
\name{Assay_Data}
\alias{Assay_Data}
\title{Assay data class}
\description{
Class of assay data providing functions for check, selection and making summary.
}
\section{Public fields}{

\describe{
  \item{\code{df_x}: }{data.frame of single-drug response data.}
  \item{\code{df_all_matrix}: }{data.frame of matrix-form observed data including points at a concentration of 0. NULL if the compound layout is not matrix-form.}
  \item{\code{df_x_matrix}: }{data.frame of matrix-form observed data. NULL if the compound layout is not matrix-form.}
  \item{\code{df_x1}: }{data.frame of single-drug response data.}
  \item{\code{df_x2}: }{data.frame of single-drug response data.}
  \item{\code{v_hc}: }{numeric vector of high control.}
  \item{\code{specimen}: }{specimen.}
  \item{\code{analysis}: }{analysis.}
  \item{\code{display_unit}: }{display_unit.}
  \item{\code{v_comp}: }{character vector of compounds.}
  \item{\code{v_metric}: }{character vector of metrics used.}
  \item{\code{data_handler}: }{Data_Handler class object that generates an input list for Stan sampling.}
}
}

\concept{Data}
\keyword{assay,}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Assay_Data-new}{\code{Assay_Data$new()}}
\item \href{#method-Assay_Data-process_data}{\code{Assay_Data$process_data()}}
\item \href{#method-Assay_Data-get_drc_input_data}{\code{Assay_Data$get_drc_input_data()}}
\item \href{#method-Assay_Data-get_bayesyc_input_data}{\code{Assay_Data$get_bayesyc_input_data()}}
\item \href{#method-Assay_Data-set_summary_data}{\code{Assay_Data$set_summary_data()}}
\item \href{#method-Assay_Data-get_meta}{\code{Assay_Data$get_meta()}}
\item \href{#method-Assay_Data-get_judgement}{\code{Assay_Data$get_judgement()}}
\item \href{#method-Assay_Data-get_lod_info}{\code{Assay_Data$get_lod_info()}}
\item \href{#method-Assay_Data-clone}{\code{Assay_Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-new"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-new}{}}}
\subsection{Method \code{new()}}{
Initialize fields by \code{data} that contains elements correspond to the fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{list of meta data and observed data corresponding to one assay.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-process_data"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-process_data}{}}}
\subsection{Method \code{process_data()}}{
Process data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$process_data(select, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{select}}{"target" for single target metric OR "multi" for multimetric.}

\item{\code{params}}{list object of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-get_drc_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-get_drc_input_data}{}}}
\subsection{Method \code{get_drc_input_data()}}{
Generate and return an input list for Stan sampling of logistic curve.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$get_drc_input_data(comp, params, metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp}}{target compound.}

\item{\code{params}}{list object of parameters.}

\item{\code{metric}}{target metric in multi-metric mode. NULL for single-metric mode.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-get_bayesyc_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-get_bayesyc_input_data}{}}}
\subsection{Method \code{get_bayesyc_input_data()}}{
Generate and return an input list for Stan sampling of BayeSyC.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$get_bayesyc_input_data(params, metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{list object of parameters.}

\item{\code{metric}}{target metric in multi-metric mode. NULL for single-metric mode.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-set_summary_data"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-set_summary_data}{}}}
\subsection{Method \code{set_summary_data()}}{
Calculate summary statistics for high and low control values of \code{v_metric}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$set_summary_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-get_meta"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-get_meta}{}}}
\subsection{Method \code{get_meta()}}{
Return meta data for stats output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$get_meta()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-get_judgement"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-get_judgement}{}}}
\subsection{Method \code{get_judgement()}}{
Pseudo function returning NULL for stats output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$get_judgement()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of judgement information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-get_lod_info"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-get_lod_info}{}}}
\subsection{Method \code{get_lod_info()}}{
Pseudo function returning NULL for stats output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$get_lod_info(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{data.frame of one particular drug response.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of limit of detection information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay_Data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Assay_Data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay_Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
