% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{cat_file_info}
\alias{cat_file_info}
\title{Concatenate file info}
\usage{
cat_file_info(i, name)
}
\arguments{
\item{v_i}{a vector of file numbers}

\item{v_name}{a vector of file names}
}
\value{
A vector of concatenated file info
}
\description{
Concatenate file info
}
