% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_data.R
\docType{class}
\name{Data_Handler}
\alias{Data_Handler}
\title{Data handler interface}
\description{
Interface of data handling for Stan input.
}
\section{Public fields}{

}

\keyword{Stan}
\keyword{data,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Data_Handler-get_drc_input_data}{\code{Data_Handler$get_drc_input_data()}}
\item \href{#method-Data_Handler-get_bayesyc_input_data}{\code{Data_Handler$get_bayesyc_input_data()}}
\item \href{#method-Data_Handler-clone}{\code{Data_Handler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data_Handler-get_drc_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Data_Handler-get_drc_input_data}{}}}
\subsection{Method \code{get_drc_input_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data_Handler$get_drc_input_data(assay_data, comp, params, metric)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data_Handler-get_bayesyc_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Data_Handler-get_bayesyc_input_data}{}}}
\subsection{Method \code{get_bayesyc_input_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data_Handler$get_bayesyc_input_data(assay_data, params, metric)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data_Handler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Data_Handler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data_Handler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
