% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_data.R
\docType{class}
\name{Data_Handler_1}
\alias{Data_Handler_1}
\title{Data handler class 1}
\description{
Concrete class of Data_Handler.
}
\section{Inherit}{

\describe{
  \item{\code{Data_Handler}}{}
}
}

\section{Public fields}{

}

\keyword{Stan}
\keyword{data,}
\section{Super class}{
\code{\link[bayesyc:Data_Handler]{bayesyc::Data_Handler}} -> \code{Data_Handler_1}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Data_Handler_1-get_drc_input_data}{\code{Data_Handler_1$get_drc_input_data()}}
\item \href{#method-Data_Handler_1-get_bayesyc_input_data}{\code{Data_Handler_1$get_bayesyc_input_data()}}
\item \href{#method-Data_Handler_1-clone}{\code{Data_Handler_1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data_Handler_1-get_drc_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Data_Handler_1-get_drc_input_data}{}}}
\subsection{Method \code{get_drc_input_data()}}{
Generate and return an input list object for logistic curve.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data_Handler_1$get_drc_input_data(assay_data, comp, params, metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assay_data}}{Assay_Data class object.}

\item{\code{comp}}{target compound.}

\item{\code{params}}{list object of parameters.}

\item{\code{metric}}{target metric in multi-metric mode. NULL for single-metric mode.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list object.
Generate and return an input list object for BayeSyC.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data_Handler_1-get_bayesyc_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Data_Handler_1-get_bayesyc_input_data}{}}}
\subsection{Method \code{get_bayesyc_input_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data_Handler_1$get_bayesyc_input_data(assay_data, params, metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assay_data}}{Assay_Data class object.}

\item{\code{params}}{list object of parameters.}

\item{\code{metric}}{target metric in multi-metric mode. NULL for single-metric mode.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data_Handler_1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Data_Handler_1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data_Handler_1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
