% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Estimates}
\alias{Estimates}
\title{Basic Estimates class}
\description{
This class behaves as an abstract class as \code{est_params} is defined by subclass.
This class holds \code{df} that contains estimated values.
}
\section{Public fields}{

\describe{
  \item{\code{est_params}: }{parameter names to set \code{df}.}
  \item{\code{df}: }{data.frame containing estimated values.
  Rows are \code{est_params} and columns are c("2.5\%", "50\%", "97.5\%").}
}
}

\seealso{
Other Base: 
\code{\link{Fit_Info}},
\code{\link{Predictions}},
\code{\link{Stats}}
}
\concept{Base}
\keyword{Stan,}
\keyword{base}
\keyword{estimates,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Estimates-set_estimates}{\code{Estimates$set_estimates()}}
\item \href{#method-Estimates-get_valid_params}{\code{Estimates$get_valid_params()}}
\item \href{#method-Estimates-check_ic50}{\code{Estimates$check_ic50()}}
\item \href{#method-Estimates-clone}{\code{Estimates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Estimates-set_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-Estimates-set_estimates}{}}}
\subsection{Method \code{set_estimates()}}{
Set \code{df} containing estimated values.
Assumed to be called from subclass with the argument \code{fit_summary}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimates$set_estimates(fit_summary)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_summary}}{matrix of summary stats of estimated parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Estimates-get_valid_params"></a>}}
\if{latex}{\out{\hypertarget{method-Estimates-get_valid_params}{}}}
\subsection{Method \code{get_valid_params()}}{
This method check the existence of \code{params_to_check} in \code{fit_summary},
and return valid params
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimates$get_valid_params(fit_summary, params_to_check)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_summary}}{matrix of summary stats of estimated parameters.}

\item{\code{params_to_check}}{parameters to be checked.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Estimates-check_ic50"></a>}}
\if{latex}{\out{\hypertarget{method-Estimates-check_ic50}{}}}
\subsection{Method \code{check_ic50()}}{
Check IC50
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimates$check_ic50(input, thre_ic50)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{input list object used for MCMC sampling.}

\item{\code{thre_ic50}}{numeric threshold for concentration range.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Estimates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Estimates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
