% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Estimates_MonoPhasic4}
\alias{Estimates_MonoPhasic4}
\title{Estimates class for monophasic logistic curve model}
\description{
Concrete class of Estimates for monophasic logistic curve model.
}
\section{Inherit}{

\describe{
  \item{\code{Estimates}}{}
}
}

\section{Public fields}{

\describe{
  \item{\code{phase}: }{number of phases in logistic curve. Set "mono" in this class.}
  \item{\code{representative}: }{representative phase in multiphasic model. Set NA in this class.}
}
}

\seealso{
Other Logistic4: 
\code{\link{Estimates_BiPhasic4}},
\code{\link{Fit_BiPhasic4}},
\code{\link{Fit_Dual}},
\code{\link{Fit_MonoPhasic4}},
\code{\link{Predictions_Curve}},
\code{\link{Stats_Logistic4}}
}
\concept{Logistic4}
\keyword{Stan}
\keyword{estimates,}
\section{Super class}{
\code{\link[bayesyc:Estimates]{bayesyc::Estimates}} -> \code{Estimates_MonoPhasic4}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Estimates_MonoPhasic4-new}{\code{Estimates_MonoPhasic4$new()}}
\item \href{#method-Estimates_MonoPhasic4-clone}{\code{Estimates_MonoPhasic4$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Estimates" data-id="check_ic50"><a href='../../bayesyc/html/Estimates.html#method-Estimates-check_ic50'><code>bayesyc::Estimates$check_ic50()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Estimates" data-id="get_valid_params"><a href='../../bayesyc/html/Estimates.html#method-Estimates-get_valid_params'><code>bayesyc::Estimates$get_valid_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Estimates" data-id="set_estimates"><a href='../../bayesyc/html/Estimates.html#method-Estimates-set_estimates'><code>bayesyc::Estimates$set_estimates()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Estimates_MonoPhasic4-new"></a>}}
\if{latex}{\out{\hypertarget{method-Estimates_MonoPhasic4-new}{}}}
\subsection{Method \code{new()}}{
Set \code{est_params} and \code{df} declared in the superclass 
using the argument \code{fit_info}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimates_MonoPhasic4$new(fit_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_info}}{Fit_info class object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Estimates_MonoPhasic4-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Estimates_MonoPhasic4-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimates_MonoPhasic4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
