% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Fit_Dual}
\alias{Fit_Dual}
\title{Fit_Dual fit information class}
\description{
Concrete class of Fit_Info for Fit_Dual fit.
}
\section{Inherit}{

\describe{
  \item{\code{Fit_Info}}{}
}
}

\section{Public fields}{

}

\seealso{
Other Logistic4: 
\code{\link{Estimates_BiPhasic4}},
\code{\link{Estimates_MonoPhasic4}},
\code{\link{Fit_BiPhasic4}},
\code{\link{Fit_MonoPhasic4}},
\code{\link{Predictions_Curve}},
\code{\link{Stats_Logistic4}}
}
\concept{Logistic4}
\keyword{Stan}
\keyword{fit,}
\section{Super class}{
\code{\link[bayesyc:Fit_Info]{bayesyc::Fit_Info}} -> \code{Fit_Dual}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Fit_Dual-new}{\code{Fit_Dual$new()}}
\item \href{#method-Fit_Dual-get_bayesyc_params}{\code{Fit_Dual$get_bayesyc_params()}}
\item \href{#method-Fit_Dual-clone}{\code{Fit_Dual$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="calculate_IC"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-calculate_IC'><code>bayesyc::Fit_Info$calculate_IC()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="check_fit"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-check_fit'><code>bayesyc::Fit_Info$check_fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="check_ic50"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-check_ic50'><code>bayesyc::Fit_Info$check_ic50()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="check_tc"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-check_tc'><code>bayesyc::Fit_Info$check_tc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="get_draws"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-get_draws'><code>bayesyc::Fit_Info$get_draws()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="get_judgement"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-get_judgement'><code>bayesyc::Fit_Info$get_judgement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="get_pi_names"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-get_pi_names'><code>bayesyc::Fit_Info$get_pi_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="select_fit_info"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-select_fit_info'><code>bayesyc::Fit_Info$select_fit_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Fit_Info" data-id="set_sf_tc"><a href='../../bayesyc/html/Fit_Info.html#method-Fit_Info-set_sf_tc'><code>bayesyc::Fit_Info$set_sf_tc()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Dual-new"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Dual-new}{}}}
\subsection{Method \code{new()}}{
Run \code{initialize} declared in the superclass.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Dual$new(fit, input, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit}}{fitting result object.}

\item{\code{input}}{list used as a input of Stan sampling.}

\item{\code{model}}{model used for fitting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Dual-get_bayesyc_params"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Dual-get_bayesyc_params}{}}}
\subsection{Method \code{get_bayesyc_params()}}{
Return estimated parameters from Posterior distributions of Dual for BayeSyC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Dual$get_bayesyc_params()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Dual-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Dual-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Dual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
