% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Fit_Info}
\alias{Fit_Info}
\title{Basic Stan fit information class}
\description{
This class behaves as an abstract class as \code{estimates}, \code{predictions} and \code{stats} are defined by subclass,
using 'Bridge pattern'.  
This class holds Stan fitting result object, input list and other basic information of Stan fit result.
}
\section{}{
NA
}

\seealso{
Other Base: 
\code{\link{Estimates}},
\code{\link{Predictions}},
\code{\link{Stats}}
}
\concept{Base}
\keyword{Stan,}
\keyword{base}
\keyword{fit,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Fit_Info-new}{\code{Fit_Info$new()}}
\item \href{#method-Fit_Info-set_sf_tc}{\code{Fit_Info$set_sf_tc()}}
\item \href{#method-Fit_Info-check_fit}{\code{Fit_Info$check_fit()}}
\item \href{#method-Fit_Info-check_tc}{\code{Fit_Info$check_tc()}}
\item \href{#method-Fit_Info-check_ic50}{\code{Fit_Info$check_ic50()}}
\item \href{#method-Fit_Info-get_pi_names}{\code{Fit_Info$get_pi_names()}}
\item \href{#method-Fit_Info-calculate_IC}{\code{Fit_Info$calculate_IC()}}
\item \href{#method-Fit_Info-get_judgement}{\code{Fit_Info$get_judgement()}}
\item \href{#method-Fit_Info-select_fit_info}{\code{Fit_Info$select_fit_info()}}
\item \href{#method-Fit_Info-get_draws}{\code{Fit_Info$get_draws()}}
\item \href{#method-Fit_Info-clone}{\code{Fit_Info$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-new"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-new}{}}}
\subsection{Method \code{new()}}{
Set \code{fit} and \code{input}, \code{fit_summary}, and \code{sf_tc} 
using information from \code{fit}.
\code{sf_tc} is set to 100 / median of estimated u_max as a default.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$new(fit, input, model, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit}}{fitting result object.}

\item{\code{input}}{list of input data used for Stan sampling.}

\item{\code{model}}{model used for fitting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-set_sf_tc"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-set_sf_tc}{}}}
\subsection{Method \code{set_sf_tc()}}{
Set \code{sf_tc} (scaling factor of T/C).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$set_sf_tc(sf_tc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-check_fit"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-check_fit}{}}}
\subsection{Method \code{check_fit()}}{
Check success of fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$check_fit(thre_rhat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thre_rhat}}{threshold for Rhat}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-check_tc"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-check_tc}{}}}
\subsection{Method \code{check_tc()}}{
Set \code{thre_tc} and \code{$ok_TC}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$check_tc(thre_tc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thre_tc}}{threshold for minimum T/C.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-check_ic50"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-check_ic50}{}}}
\subsection{Method \code{check_ic50()}}{
Set \code{thre_tc} and \code{$ok_TC}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$check_ic50(thre_ic50)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thre_tc}}{threshold for detected IC50.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-get_pi_names"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-get_pi_names}{}}}
\subsection{Method \code{get_pi_names()}}{
Return pi[] variable names if estimated.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$get_pi_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character "pi[]"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-calculate_IC"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-calculate_IC}{}}}
\subsection{Method \code{calculate_IC()}}{
Calculate and set \code{waic}, \code{looic} and return Estimate ± SE of them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$calculate_IC()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-get_judgement"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-get_judgement}{}}}
\subsection{Method \code{get_judgement()}}{
Return c(\code{success_fit}, \code{ok_Rhat}, \code{ok_TC}, \code{ok_IC50}) for stats.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$get_judgement()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-select_fit_info"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-select_fit_info}{}}}
\subsection{Method \code{select_fit_info()}}{
Choose one Fit_Info object from \code{self} and \code{fit_info_alt} by comparing their WAIC.
Myself (\code{self}) is the primary choice if alt_first=FALSE (default).
Alternative (\code{fit_info_alt}) is selected if it satisfies the criteria.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$select_fit_info(fit_info_alt, thre_rhat, alt_first = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_info_alt}}{an alternative Fit_Info object.}

\item{\code{thre_rhat}}{threshold for Rhat}

\item{\code{alt_first}}{FLASE (default). TRUE to prioritize alternative Fit_Info object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A selected Fit_Info object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-get_draws"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-get_draws}{}}}
\subsection{Method \code{get_draws()}}{
Return a matrix of MCMC draws.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$get_draws()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
matrix of MCMC draws
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Fit_Info-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Fit_Info-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fit_Info$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
