% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_beta_param}
\alias{get_beta_param}
\title{Get Beta distribution parameters from MCMC samples}
\usage{
get_beta_param(mx_samples, rv, coef_var = 1)
}
\arguments{
\item{mx_samples}{Matrix of MCMC samples}

\item{rv}{random variable of interest}

\item{coef_var}{coefficient applied to variance}
}
\value{
A vector of beta distribution parameters, alpha and beta
}
\description{
Get Beta distribution parameters from MCMC samples
}
