% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{get_fig_title}
\alias{get_fig_title}
\title{Get figure title}
\usage{
get_fig_title(specimen, display_type, du)
}
\arguments{
\item{specimen}{specimen}

\item{display_type}{display unit type}

\item{du}{display unit}
}
\value{
Title
}
\description{
Get figure title
}
