% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{get_second_legend_value}
\alias{get_second_legend_value}
\title{Get legend value for second key}
\usage{
get_second_legend_value(df, v_def_value, first_key, second_key)
}
\arguments{
\item{df}{data.frame used for plot}

\item{v_def_value}{definition of value used for manual setting}

\item{first_key}{first key}

\item{second_key}{second key}
}
\value{
v_value legend value for second key
}
\description{
Get legend value for second key
}
