% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{class}
\name{Manual_Input_Double}
\alias{Manual_Input_Double}
\title{Class for loading two-drug combination dose response data}
\description{
This class for manually loading two-drug combination dose response data
}
\section{Public fields}{

\describe{
  \item{\code{ls_data}: }{list of Assay_Data class objects.}
  \item{\code{v_specimen}: }{character vector of specimen IDs.}
  \item{\code{v_analysis}: }{character vector of analysis IDs.}
  \item{\code{uniq_display_units}: }{character vector of unique display units.}
}
}

\seealso{
Other Input: 
\code{\link{Manual_Input_Single}},
\code{\link{Whole_Input_ONeil}}
}
\concept{Input}
\keyword{double}
\keyword{manual,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Manual_Input_Double-new}{\code{Manual_Input_Double$new()}}
\item \href{#method-Manual_Input_Double-process_input_data}{\code{Manual_Input_Double$process_input_data()}}
\item \href{#method-Manual_Input_Double-clone}{\code{Manual_Input_Double$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Manual_Input_Double-new"></a>}}
\if{latex}{\out{\hypertarget{method-Manual_Input_Double-new}{}}}
\subsection{Method \code{new()}}{
Load input data.frame and standardize data format for downstream analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Manual_Input_Double$new(params, df_matrix, base_value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{list object of parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Manual_Input_Double-process_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Manual_Input_Double-process_input_data}{}}}
\subsection{Method \code{process_input_data()}}{
Process input data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Manual_Input_Double$process_input_data(select, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{select}}{what to be selected: "target" or "multi"}

\item{\code{params}}{list object of parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of error messages. NULL if no error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Manual_Input_Double-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Manual_Input_Double-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Manual_Input_Double$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
