% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{class}
\name{Manual_Input_Single}
\alias{Manual_Input_Single}
\title{Class for loading single-drug dose response data}
\description{
This class for manually loading single-drug dose response data
}
\section{Public fields}{

\describe{
  \item{\code{ls_data}: }{list of Assay_Data class objects.}
  \item{\code{v_specimen}: }{character vector of specimen IDs.}
  \item{\code{v_analysis}: }{character vector of analysis IDs.}
  \item{\code{uniq_specimens}: }{character vector of unique specimen IDs.}
  \item{\code{uniq_compounds}: }{character vector of unique compounds.}
  \item{\code{n_uniq_specimens}: }{number of unique specimen IDs.}
  \item{\code{n_uniq_compounds}: }{number of unique compounds.}
  \item{\code{uniq_metrics}: }{character vector of unique metrics.}
  \item{\code{n_uniq_metrics}: }{number of unique metrics.}
  \item{\code{uniq_display_units}: }{character vector of unique display units.}
}
}

\seealso{
Other Input: 
\code{\link{Manual_Input_Double}},
\code{\link{Whole_Input_ONeil}}
}
\concept{Input}
\keyword{manual,}
\keyword{single}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Manual_Input_Single-new}{\code{Manual_Input_Single$new()}}
\item \href{#method-Manual_Input_Single-process_input_data}{\code{Manual_Input_Single$process_input_data()}}
\item \href{#method-Manual_Input_Single-clone}{\code{Manual_Input_Single$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Manual_Input_Single-new"></a>}}
\if{latex}{\out{\hypertarget{method-Manual_Input_Single-new}{}}}
\subsection{Method \code{new()}}{
Load input data.frame and standardize data format for downstream analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Manual_Input_Single$new(params, df_single, base_value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{list object of parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Manual_Input_Single-process_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Manual_Input_Single-process_input_data}{}}}
\subsection{Method \code{process_input_data()}}{
Process input data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Manual_Input_Single$process_input_data(select, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{select}}{what to be selected: "target" or "multi"}

\item{\code{params}}{list object of parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of error messages. NULL if no error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Manual_Input_Single-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Manual_Input_Single-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Manual_Input_Single$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
