% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{modify_ggplotly_legend}
\alias{modify_ggplotly_legend}
\title{Modify ggplotly legend}
\usage{
modify_ggplotly_legend(obj_ggplotly)
}
\arguments{
\item{obj_ggplotly}{ggplotly object}
}
\value{
obj_ggplotly ggplotly object
}
\description{
Remove unnecessary parentheses "(*****,1)" and put space after comma
}
