% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Predictions}
\alias{Predictions}
\title{Basic Predictions class}
\description{
This class behaves as an abstract class as \code{y_pred_names} and \code{u_pred_names} 
are defined by subclass.
This class holds \code{df} that contains predicted values of \code{y_pred_names} and \code{u_pred_names}.
}
\section{Public fields}{

\describe{
  \item{\code{y_pred_names}: }{parameter names of predicted y to set \code{df}.}
  \item{\code{u_pred_names}: }{parameter names of predicted u to set \code{df}.}
  \item{\code{df}: }{data.frame containing predicted values that are 2.5, 50 and 97.5 percentile values 
  of \code{y_pred_names}, \code{u_pred_names} and their T/C values.}
  \item{\code{xx}: }{vector of new x values used in the prediction.}
}
}

\seealso{
Other Base: 
\code{\link{Estimates}},
\code{\link{Fit_Info}},
\code{\link{Stats}}
}
\concept{Base}
\keyword{Stan,}
\keyword{base}
\keyword{predictions,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Predictions-set_predictions}{\code{Predictions$set_predictions()}}
\item \href{#method-Predictions-check_tc}{\code{Predictions$check_tc()}}
\item \href{#method-Predictions-clone}{\code{Predictions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Predictions-set_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-Predictions-set_predictions}{}}}
\subsection{Method \code{set_predictions()}}{
Set of predicted 2.5, 50 and 97.5 percentile values of y, u, tc_y and tc_u.
Assumed to be called from subclass.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictions$set_predictions(fit_summary, input, sf_tc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_summary}}{matrix of summary stats of estimated parameters.}

\item{\code{input}}{list used as a input of Stan sampling.}

\item{\code{sf_tc}}{numeric scaling factor to calculate T/C.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Predictions-check_tc"></a>}}
\if{latex}{\out{\hypertarget{method-Predictions-check_tc}{}}}
\subsection{Method \code{check_tc()}}{
Return TRUE if any median T/Cs > \code{thre_tc} otherwise FALSE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictions$check_tc(thre_tc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thre_tc}}{\% threshold for T/C}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if minimum of T/C > \code{thre_tc} otherwise FALSE.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Predictions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Predictions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
