% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Predictions_Curve}
\alias{Predictions_Curve}
\title{Predictions class for curve values}
\description{
Concrete class of Predictions for curve values.
}
\section{Inherit}{

\describe{
  \item{\code{Predictions}}{}
}
}

\section{Public fields}{

}

\seealso{
Other Logistic4: 
\code{\link{Estimates_BiPhasic4}},
\code{\link{Estimates_MonoPhasic4}},
\code{\link{Fit_BiPhasic4}},
\code{\link{Fit_Dual}},
\code{\link{Fit_MonoPhasic4}},
\code{\link{Stats_Logistic4}}
}
\concept{Logistic4}
\keyword{Stan}
\keyword{predictions,}
\section{Super class}{
\code{\link[bayesyc:Predictions]{bayesyc::Predictions}} -> \code{Predictions_Curve}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Predictions_Curve-new}{\code{Predictions_Curve$new()}}
\item \href{#method-Predictions_Curve-clone}{\code{Predictions_Curve$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Predictions" data-id="check_tc"><a href='../../bayesyc/html/Predictions.html#method-Predictions-check_tc'><code>bayesyc::Predictions$check_tc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Predictions" data-id="set_predictions"><a href='../../bayesyc/html/Predictions.html#method-Predictions-set_predictions'><code>bayesyc::Predictions$set_predictions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Predictions_Curve-new"></a>}}
\if{latex}{\out{\hypertarget{method-Predictions_Curve-new}{}}}
\subsection{Method \code{new()}}{
Set \code{y_pred_names}, \code{u_pred_names}, \code{df} and \code{xx} declared in the superclass
using the argument \code{fit_info}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictions_Curve$new(fit_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_info}}{Fit_info class object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Predictions_Curve-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Predictions_Curve-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictions_Curve$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
