% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Predictions_Surface}
\alias{Predictions_Surface}
\title{Predictions class for surface values}
\description{
Concrete class of Predictions for surface values.
}
\section{Inherit}{

\describe{
  \item{\code{Predictions}}{}
}
}

\section{Public fields}{

\describe{
  \item{\code{mx_surface_median}: }{matrix form of predicted u at 50 percentile.}
  \item{\code{mx_surface_lower}: }{matrix form of predicted u at 2.5 percentile.}
  \item{\code{mx_surface_upper}: }{matrix form of predicted u at 97.5 percentile.}
  \item{\code{mx_surface_median_tc}: }{matrix form of predicted T/C u at 50 percentile.}
}
}

\seealso{
Other BayeSyC4: 
\code{\link{Estimates_BayeSyC4}},
\code{\link{Fit_BayeSyC4}},
\code{\link{Stats_BayeSyC4}}
}
\concept{BayeSyC4}
\keyword{Stan}
\keyword{predictions,}
\section{Super class}{
\code{\link[bayesyc:Predictions]{bayesyc::Predictions}} -> \code{Predictions_Surface}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Predictions_Surface-new}{\code{Predictions_Surface$new()}}
\item \href{#method-Predictions_Surface-clone}{\code{Predictions_Surface$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Predictions" data-id="check_tc"><a href='../../bayesyc/html/Predictions.html#method-Predictions-check_tc'><code>bayesyc::Predictions$check_tc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Predictions" data-id="set_predictions"><a href='../../bayesyc/html/Predictions.html#method-Predictions-set_predictions'><code>bayesyc::Predictions$set_predictions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Predictions_Surface-new"></a>}}
\if{latex}{\out{\hypertarget{method-Predictions_Surface-new}{}}}
\subsection{Method \code{new()}}{
Set \code{y_pred_names}, \code{u_pred_names} and \code{df} declared in the superclass
using the argument \code{fit_info}.
Set \code{mx_surface_median}, \code{mx_surface_lower}, \code{mx_surface_upper} and \code{mx_surface_median_tc}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictions_Surface$new(fit_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_info}}{Fit_info class object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Predictions_Surface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Predictions_Surface-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictions_Surface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
