% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Stats}
\alias{Stats}
\title{Basic Stats class}
\description{
This class behaves as an abstract class as \code{stats_params} is defined by subclass.
This class holds stats vector \code{values} that contains information of 
\code{stats_param} renamed as \code{stats_name}.
}
\section{Public fields}{

\describe{
  \item{\code{stats_params}: }{parameter names to construct stats vector \code{values}.}
  \item{\code{values}: }{vector of stats that are 2.5\%, 50\%, 97.5\%, Rhat and Neff values of \code{stats_params}.}
}
}

\seealso{
Other Base: 
\code{\link{Estimates}},
\code{\link{Fit_Info}},
\code{\link{Predictions}}
}
\concept{Base}
\keyword{Stan,}
\keyword{base}
\keyword{estimates,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Stats-set_stats}{\code{Stats$set_stats()}}
\item \href{#method-Stats-clone}{\code{Stats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stats-set_stats"></a>}}
\if{latex}{\out{\hypertarget{method-Stats-set_stats}{}}}
\subsection{Method \code{set_stats()}}{
Set stats vector \code{values}.
Assumed to be called from subclass with the argument \code{fit_summary} and \code{estimates}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stats$set_stats(fit_summary, estimates, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_summary}}{matrix of summary stats of estimated parameters.}

\item{\code{estimates}}{Estimates class object.}

\item{\code{model}}{model used for fitting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stats-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Stats-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
