% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\docType{class}
\name{Stats_BayeSyC4}
\alias{Stats_BayeSyC4}
\title{Stats class for BayeSyC model}
\description{
Concrete class of Stats for BayeSyC model.
}
\section{Inherit}{

\describe{
  \item{\code{Stats}}{}
}
}

\section{Public fields}{

}

\seealso{
Other BayeSyC4: 
\code{\link{Estimates_BayeSyC4}},
\code{\link{Fit_BayeSyC4}},
\code{\link{Predictions_Surface}}
}
\concept{BayeSyC4}
\keyword{Stan}
\keyword{stats,}
\section{Super class}{
\code{\link[bayesyc:Stats]{bayesyc::Stats}} -> \code{Stats_BayeSyC4}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Stats_BayeSyC4-new}{\code{Stats_BayeSyC4$new()}}
\item \href{#method-Stats_BayeSyC4-clone}{\code{Stats_BayeSyC4$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Stats" data-id="set_stats"><a href='../../bayesyc/html/Stats.html#method-Stats-set_stats'><code>bayesyc::Stats$set_stats()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stats_BayeSyC4-new"></a>}}
\if{latex}{\out{\hypertarget{method-Stats_BayeSyC4-new}{}}}
\subsection{Method \code{new()}}{
Set \code{stats_params} and \code{values} declared in the superclass 
using the argument \code{fit_info}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stats_BayeSyC4$new(fit_info)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_info}}{Fit_info class object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stats_BayeSyC4-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Stats_BayeSyC4-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stats_BayeSyC4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
