% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{class}
\name{Whole_Input_ONeil}
\alias{Whole_Input_ONeil}
\title{Class for loading ONeil data}
\description{
This class handle O'Neil et al. 2016 data.
}
\section{Public fields}{

\describe{
  \item{\code{ls_data}: }{list of Assay_Data class objects.}
  \item{\code{v_specimen}: }{character vector of specimen IDs.}
  \item{\code{v_analysis}: }{character vector of analysis IDs.}
  \item{\code{n_uniq_metrics}: }{number of unique metrics.}
}
}

\seealso{
Other Input: 
\code{\link{Manual_Input_Double}},
\code{\link{Manual_Input_Single}}
}
\concept{Input}
\keyword{ONeil,}
\keyword{double}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Whole_Input_ONeil-new}{\code{Whole_Input_ONeil$new()}}
\item \href{#method-Whole_Input_ONeil-process_input_data}{\code{Whole_Input_ONeil$process_input_data()}}
\item \href{#method-Whole_Input_ONeil-clone}{\code{Whole_Input_ONeil$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Input_ONeil-new"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Input_ONeil-new}{}}}
\subsection{Method \code{new()}}{
Load input data.frame and standardize data format for downstream analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Input_ONeil$new(params, df_double, df_single)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{list object of parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Input_ONeil-process_input_data"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Input_ONeil-process_input_data}{}}}
\subsection{Method \code{process_input_data()}}{
Process input data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Input_ONeil$process_input_data(select, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{select}}{what to be selected: "target" or "multi"}

\item{\code{params}}{list object of parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of error messages. NULL if no error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Input_ONeil-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Input_ONeil-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Input_ONeil$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
