% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{class}
\name{Whole_Output}
\alias{Whole_Output}
\title{Class for all fitting results}
\description{
Class for all fitting results. This class should be inherited by concrete class.
}
\section{Public fields}{

\describe{
  \item{\code{ls_data}: }{list of Plate_Data class objects.}
  \item{\code{params}: }{knit param list.}
  \item{\code{v_n_fit}: }{character of plate ID.}
}
}

\seealso{
Other Output: 
\code{\link{Whole_Output_BayeSyC}},
\code{\link{Whole_Output_Multi}}
}
\concept{Output}
\keyword{Stan}
\keyword{fit,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Whole_Output-execute_stan_sampling}{\code{Whole_Output$execute_stan_sampling()}}
\item \href{#method-Whole_Output-clone}{\code{Whole_Output$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output-execute_stan_sampling"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output-execute_stan_sampling}{}}}
\subsection{Method \code{execute_stan_sampling()}}{
Execute MCMC sampling. If the result does not converge, it re-runs \code{self$params$n_retry} times.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output$execute_stan_sampling(stan_model, ls_in, id, mcmc_timeout = 20000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stan_model}}{model (stanmodel or CmdStanModel) object.}

\item{\code{ls_in}}{list object of input data.}

\item{\code{id}}{ID of sampling}

\item{\code{mcmc_timeout}}{max CPU time (time used for all cores are summed up) for MCMC sampling}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
