% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{class}
\name{Whole_Output_BayeSyC}
\alias{Whole_Output_BayeSyC}
\title{Class for all BayeSyC fitting results}
\description{
This class conduct BayeSyC fitting and processing of the results for output.
}
\section{Inherit}{

\describe{
  \item{\code{Whole_Output}}{}
}
}

\section{Public fields}{

\describe{
  \item{\code{stan_model_drc}: }{model (stanmodel or CmdStanModel) object of mono-phasic logistic curve.}
  \item{\code{stan_model_biphasic}: }{model (stanmodel or CmdStanModel) object of bi-phasic logistic curve.}
  \item{\code{ls_fit_bayesyc}: }{list of Fit_Info class objects.}
  \item{\code{ls_res_bayesyc}: }{list of processed fitting results.}
  \item{\code{df_stats_bayesyc}: }{data.frame of stats.}
}
}

\seealso{
Other Output: 
\code{\link{Whole_Output}},
\code{\link{Whole_Output_Multi}}
}
\concept{Output}
\keyword{Stan}
\keyword{fit,}
\section{Super class}{
\code{\link[bayesyc:Whole_Output]{bayesyc::Whole_Output}} -> \code{Whole_Output_BayeSyC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Whole_Output_BayeSyC-new}{\code{Whole_Output_BayeSyC$new()}}
\item \href{#method-Whole_Output_BayeSyC-perform_fitting_bayesyc}{\code{Whole_Output_BayeSyC$perform_fitting_bayesyc()}}
\item \href{#method-Whole_Output_BayeSyC-perform_fitting_bayesyc_1}{\code{Whole_Output_BayeSyC$perform_fitting_bayesyc_1()}}
\item \href{#method-Whole_Output_BayeSyC-perform_fitting_bayesyc_2}{\code{Whole_Output_BayeSyC$perform_fitting_bayesyc_2()}}
\item \href{#method-Whole_Output_BayeSyC-process_fit_bayesyc}{\code{Whole_Output_BayeSyC$process_fit_bayesyc()}}
\item \href{#method-Whole_Output_BayeSyC-make_stats_bayesyc}{\code{Whole_Output_BayeSyC$make_stats_bayesyc()}}
\item \href{#method-Whole_Output_BayeSyC-clone}{\code{Whole_Output_BayeSyC$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Whole_Output" data-id="execute_stan_sampling"><a href='../../bayesyc/html/Whole_Output.html#method-Whole_Output-execute_stan_sampling'><code>bayesyc::Whole_Output$execute_stan_sampling()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_BayeSyC-new"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_BayeSyC-new}{}}}
\subsection{Method \code{new()}}{
Run sampling and process the fitting results for output.
If argument \code{$run_all} FALSE, just making an instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_BayeSyC$new(
  whole_input,
  stan_model_dual,
  stan_model_bayesyc,
  params,
  run_all = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{whole_input}}{Whole_Input class object.}

\item{\code{stan_model_dual}}{model (stanmodel or CmdStanModel) object of dual mono-phasic logistic curves.}

\item{\code{stan_model_bayesyc}}{model (stanmodel or CmdStanModel) object of BayeSyC response surface.}

\item{\code{params}}{list object of parameters}

\item{\code{run_all}}{TRUE (default). FALSE for not running analyses and processes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_BayeSyC-perform_fitting_bayesyc"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_BayeSyC-perform_fitting_bayesyc}{}}}
\subsection{Method \code{perform_fitting_bayesyc()}}{
Run perform_fitting_multi_bayesyc_1 or perform_fitting_multi_bayesyc_2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_BayeSyC$perform_fitting_bayesyc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_BayeSyC-perform_fitting_bayesyc_1"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_BayeSyC-perform_fitting_bayesyc_1}{}}}
\subsection{Method \code{perform_fitting_bayesyc_1()}}{
Run MCMC sampling and hold the fitting results as \code{self$ls_fit_multi_bayesyc}.
Case of no Dual.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_BayeSyC$perform_fitting_bayesyc_1()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_BayeSyC-perform_fitting_bayesyc_2"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_BayeSyC-perform_fitting_bayesyc_2}{}}}
\subsection{Method \code{perform_fitting_bayesyc_2()}}{
Run MCMC sampling and hold the fitting results as \code{self$ls_fit_multi_bayesyc}.
Case of with Dual followed by BayeSyC.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_BayeSyC$perform_fitting_bayesyc_2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_BayeSyC-process_fit_bayesyc"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_BayeSyC-process_fit_bayesyc}{}}}
\subsection{Method \code{process_fit_bayesyc()}}{
Process the fitting results and hold processed results as \code{self$ls_res_multi}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_BayeSyC$process_fit_bayesyc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_BayeSyC-make_stats_bayesyc"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_BayeSyC-make_stats_bayesyc}{}}}
\subsection{Method \code{make_stats_bayesyc()}}{
Make stats from the processed results and hold them as \code{self$df_stats_multi}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_BayeSyC$make_stats_bayesyc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_BayeSyC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_BayeSyC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_BayeSyC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
