% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{class}
\name{Whole_Output_Multi}
\alias{Whole_Output_Multi}
\title{Class for all multi-metric logistic curve fitting results}
\description{
This class conduct multi-metric logistic curve fitting and processing of the results for output.
}
\section{Inherit}{

\describe{
  \item{\code{Whole_Output}}{}
}
}

\section{Public fields}{

\describe{
  \item{\code{stan_model_drc}: }{model (stanmodel or CmdStanModel) object of mono-phasic logistic curve.}
  \item{\code{stan_model_biphasic}: }{model (stanmodel or CmdStanModel) object of bi-phasic logistic curve.}
  \item{\code{ls_fit_multi}: }{list of Fit_Info class objects.}
  \item{\code{ls_res_multi}: }{list of processed fitting results.}
  \item{\code{df_stats_multi}: }{data.frame of stats.}
  \item{\code{ls_out_multi}: }{list of output-ready data for plots.}
}
}

\seealso{
Other Output: 
\code{\link{Whole_Output}},
\code{\link{Whole_Output_BayeSyC}}
}
\concept{Output}
\keyword{Stan}
\keyword{fit,}
\section{Super class}{
\code{\link[bayesyc:Whole_Output]{bayesyc::Whole_Output}} -> \code{Whole_Output_Multi}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Whole_Output_Multi-new}{\code{Whole_Output_Multi$new()}}
\item \href{#method-Whole_Output_Multi-perform_fitting_multi}{\code{Whole_Output_Multi$perform_fitting_multi()}}
\item \href{#method-Whole_Output_Multi-process_fit_multi}{\code{Whole_Output_Multi$process_fit_multi()}}
\item \href{#method-Whole_Output_Multi-make_stats_multi}{\code{Whole_Output_Multi$make_stats_multi()}}
\item \href{#method-Whole_Output_Multi-make_plot_data_multi}{\code{Whole_Output_Multi$make_plot_data_multi()}}
\item \href{#method-Whole_Output_Multi-clone}{\code{Whole_Output_Multi$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bayesyc" data-topic="Whole_Output" data-id="execute_stan_sampling"><a href='../../bayesyc/html/Whole_Output.html#method-Whole_Output-execute_stan_sampling'><code>bayesyc::Whole_Output$execute_stan_sampling()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_Multi-new"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_Multi-new}{}}}
\subsection{Method \code{new()}}{
Run MCMC sampling and process the fitting results for output.
If argument \code{$run_all} FALSE, just making an instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_Multi$new(
  whole_input,
  stan_model_drc,
  stan_model_biphasic,
  params,
  run_all = TRUE,
  bi_first = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{whole_input}}{Whole_Input class object.}

\item{\code{stan_model_drc}}{model (stanmodel or CmdStanModel) object of mono-phasic logistic curve.}

\item{\code{stan_model_biphasic}}{model (stanmodel or CmdStanModel) object of bi-phasic logistic curve.}

\item{\code{params}}{list object of parameters}

\item{\code{run_all}}{TRUE (default). FALSE for not running analyses and processes.}

\item{\code{bi_first}}{FLASE (default). TRUE to prioritize bi-phasic logistic curve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_Multi-perform_fitting_multi"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_Multi-perform_fitting_multi}{}}}
\subsection{Method \code{perform_fitting_multi()}}{
Run MCMC sampling and hold the fitting results as \code{self$ls_fit_multi}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_Multi$perform_fitting_multi(bi_first = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bi_first}}{FLASE (default). TRUE to prioritize bi-phasic logistic curve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_Multi-process_fit_multi"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_Multi-process_fit_multi}{}}}
\subsection{Method \code{process_fit_multi()}}{
Process the fitting results and hold processed results as \code{self$ls_res_multi}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_Multi$process_fit_multi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_Multi-make_stats_multi"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_Multi-make_stats_multi}{}}}
\subsection{Method \code{make_stats_multi()}}{
Make stats from the processed results and hold them as \code{self$df_stats_multi}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_Multi$make_stats_multi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_Multi-make_plot_data_multi"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_Multi-make_plot_data_multi}{}}}
\subsection{Method \code{make_plot_data_multi()}}{
Make plot data from the processed results and hold them as \code{self$ls_out_multi}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_Multi$make_plot_data_multi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Whole_Output_Multi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Whole_Output_Multi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Whole_Output_Multi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
